+++ common/src/dynamic_amm/mod.rs
@@ -1,0 +1,5 @@
+pub mod aux_lp_mint;
+pub mod ix_account_builder;
+pub mod pda;
+
+pub use aux_lp_mint::*;