+++ common/src/dynamic_vault/mod.rs
@@ -1,0 +1,4 @@
+pub mod aux_lp_mint;
+pub mod pda;
+
+pub use aux_lp_mint::*;